Moved get_entry_description to another file so it can be used in other searches too
